Refactored input data hanlding
The element data and element data holder stuff was really unintuitive so
we now just use an input data structure which contains the data
necessary for visualization. This also makes elements more simple and
gives them the option to access any input data, so it is now possible to
create buttons that only light up if a combination of buttons are
pressed.