Plugin, Presets, CCT: Switch to SDL2 for gamepad input
This removes the old gamepad library and instead uses SDL2 for all
gamepad input. This also means that the keycodes for gamepad buttons
have changed. All existing presets have been updated and the config tool
will convert the old codes when loading any preset. So conversion is as
easy as loading a preset and then saving it again.

The client has not been updated for now as there's more refactoring
needed for that.