Fix #200: add setting to ignore repeated keys
Also add `keycode` and `isPressing` parameters to `updateUI()` and streamline it to make it easier to detect repeated keys and decide what to do with them.

Changed when new elements are added. Now it happens only when pressing a new key. That prevents a small space being added to the bottom of the history panel because of the top margin of an empty combination on release.