Input history: Fix #200: add repeat setting; Fix empty space at bottom; Align div instead of padding (#331)
* Add mouse buttons and scroll wheel support

* Fix #200: add setting to ignore repeated keys

Also add `keycode` and `isPressing` parameters to `updateUI()` and streamline it to make it easier to detect repeated keys and decide what to do with them.

Changed when new elements are added. Now it happens only when pressing a new key. That prevents a small space being added to the bottom of the history panel because of the top margin of an empty combination on release.

* Don't render icons div; Align body to bottom

The icons div was still taking invisible space. Now it's not rendered at all.

Align body to the bottom instead of forcing space into history div with padding.

* Update input history README with repeated key option